frontend/requirements.txt
streamlit>=1.28.0          # Основной дашборд
pandas>=1.5.0              # Работа с данными
numpy>=1.21.0              # Числовые операции
matplotlib>=3.5.0          # Графики
seaborn>=0.11.0            # Красивые визуализации
joblib>=1.2.0              # Загрузка сохранённой модели
pyyaml>=6.0                # Чтение config/params.yml
requests>=2.28.0           # Запросы к backend API (если будет интеграция)
pathlib2>=2.3.7            # Работа с путями (Path)
scikit-learn>=1.2.0        # Для препроцессора (если нужен transform)